chore: remote 모듈에서만 REST_API_KEY 사용
