+++ core/data/src/main/java/com/hongstudio/data/datasource/remote/DocumentRemoteDataSource.kt
@@ -5,9 +5,7 @@ import com.hongstudio.data.model.DocumentDto
 interface DocumentRemoteDataSource {
 
     suspend fun getSearchedImages(
-        authorization: String,
         query: String,
-        page: Int,
-        size: Int
+        page: Int
     ): List<DocumentDto>
 }
