+++ core/remote/src/main/java/com/hongstudio/remote/datasource/DocumentRemoteDataSourceImpl.kt
@@ -2,6 +2,7 @@ package com.hongstudio.remote.datasource
 
 import com.hongstudio.data.datasource.remote.DocumentRemoteDataSource
 import com.hongstudio.data.model.DocumentDto
+import com.hongstudio.remote.BuildConfig
 import com.hongstudio.remote.api.SearchImageApi
 import com.hongstudio.remote.model.toDto
 import javax.inject.Inject
@@ -11,20 +12,22 @@ class DocumentRemoteDataSourceImpl @Inject constructor(
 ) : DocumentRemoteDataSource {
 
     override suspend fun getSearchedImages(
-        authorization: String,
         query: String,
-        page: Int,
-        size: Int
+        page: Int
     ): List<DocumentDto> {
         return searchImageApi.getSearchedImages(
-            authorization = authorization,
+            authorization = BuildConfig.REST_API_KEY,
             query = query,
             page = page,
-            size = size
+            size = SIZE_PER_PAGE
         ).let {
             it.documents.map { documentRemote ->
                 documentRemote.toDto()
             }
         }
+    }
+
+    companion object {
+        private const val SIZE_PER_PAGE = 10
     }
 }
